handling phantom crashes more gracefully
a new event is now available, called "exit". When used, it will
attach a handler to be called in the event of an unexpected end
to the phantomjs process.

without a handler specified to the `phantom` lib, it will throw
an uncaught exception! however, if the user does _not_ specify
a handler of their own, we will simply throw an uncaught exception
ourselves. (but thankfully, the message will be more clear as to
what happened)