Remove `process` event listeners when a Nightmare instance ends.
This is one of several issues @fr- noted in #282. It is also important because the registered listener could cause the Nightmare instance to be retained even if there are no other references to it, which means instances can never be garbage collected.

This also removes the uncaught exception listener on the parent process, as mentioned in https://github.com/segmentio/nightmare/issues/586#issuecomment-212681749.

Finally, ensure the instance gets marked as ended even if the child process connection has already died (e.g. it crashed).