Add timeout to `goto` action. (Defaults to 30 seconds.) Addresses issues noted in #612, but does not solve the core problem.
Note this necessitates a much bigger change than one might expect -- Nightmare checks whether the window is ready to receive actions (i.e. whether it is loading) between every action, but since we can now be in a state where we are ok to continue but loading is still going on, we have to track for ourselves whether is OK to perform the next action instead of simply asking Electron whether resources are loading. This is accomplished by way of a method named `isReady()` that returns a boolean and `setIsReady(newValue)` that takes a boolean and fires an event on the `BrowserWindow` instance called `did-change-is-ready`.

The actual timeout value can be customized at construction time (via an option called `gotoTimeout`, analagous to `waitTimeout`) and at method time with a third argument `goto(url, headers, timeout)`.