Stop sending electron event data in forwarded events.
Electron events now have a few circular references and are not safe to JSON serialize (serializing causes exceptions in CI). We've had this happen before with other parts of the event, so this time around, rather than explicitly removing the problematic part, we are removing the raw event altogether in case we see this again (in its place we send an empty object so as not to change the signature of a listener).