Alter default behavior to non-persistent partition
The current behavior uses electron default, which creates a partition
for storage which persists between instances. This makes sense for
electron but not necessarily for primary use cases of Nightmare.

This change sets the default partition to be a non-persistent partition
but allows a user to opt back into persistence with simple configuration options.

Fixes segmentio/nightmare#916