Detach from debugger after forcing re-rendering
If we leave the remote debugger attached, users won't be able to open the local debugger in a visible window for manual testing after a screenshot has been taken (or rather, they can open it, but it won't work). To accomodate manual debugger usage, immediately detach once our work is done.